+++ src/components/days.js
@@ -1,0 +1,7 @@
+const createDaysTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+export {createDaysTemplate};
