+++ src/components/edit-event/edit-event-detail/edit-event-details.js
@@ -1,0 +1,13 @@
+import {createEditEventOffersTemplate} from './edit-event-offers';
+import {createEditEventDestinationTemplate} from './edit-event-destination';
+
+const createEditEventDetailsTemplate = () => {
+  return (
+    `<section class="event__details">
+      ${createEditEventOffersTemplate()}
+      ${createEditEventDestinationTemplate()}
+    </section>`
+  );
+};
+
+export {createEditEventDetailsTemplate};
