+++ src/components/edit-event/edit-event-header/edit-event-price.js
@@ -1,0 +1,19 @@
+const createEditEventPriceTemplate = () => {
+  return (
+    `<div class="event__field-group  event__field-group--price">
+      <label class="event__label" for="event-price-1">
+        <span class="visually-hidden">Price</span>
+        &euro;
+      </label>
+      <input
+        class="event__input  event__input--price"
+        id="event-price-1"
+        type="text"
+        name="event-price"
+        value=""
+      >
+    </div>`
+  );
+};
+
+export {createEditEventPriceTemplate};
