+++ src/components/edit-event/edit-event.js
@@ -1,13 +1,13 @@
-import {createEditEventHeaderTemplate} from './templates/edit-event-header';
-import {createEditEventDetails} from './templates/edit-event-details';
+import {createEditEventHeaderTemplate} from './edit-event-header/edit-event-header';
+import {createEditEventDetailsTemplate} from './templates/edit-event-details';
 
-const createTemplateEditEvent = (currentEvent) => {
+const createEditEventTemplate = (currentEvent) => {
   return (
     `<form class="trip-events__item  event  event--edit" action="#" method="post">
       ${createEditEventHeaderTemplate(currentEvent)}
-      ${createEditEventDetails()}
+      ${createEditEventDetailsTemplate()}
   </form>`
   );
 };
 
-export {createTemplateEditEvent};
+export {createEditEventTemplate};
