+++ src/components/trip-sort.js
@@ -9,7 +9,7 @@ const sortIcon = (
   </svg>`
 );
 
-const createTemplateTripSort = () => {
+const createTripSortTemplate = () => {
   return (
     `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
       <span class="trip-sort__item  trip-sort__item--day">
@@ -43,4 +43,4 @@ const createTemplateTripSort = () => {
   );
 };
 
-export {createTemplateTripSort};
+export {createTripSortTemplate};
