+++ src/components/trip-tabs.js
@@ -1,4 +1,4 @@
-const createTemplateTripTabs = () => {
+const createTripTabsTemplate = () => {
   return (
     `<nav class="trip-controls__trip-tabs  trip-tabs">
       <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
@@ -7,4 +7,4 @@ const createTemplateTripTabs = () => {
   );
 };
 
-export {createTemplateTripTabs};
+export {createTripTabsTemplate};
